+++ crates/symbolicator-js/src/interface.rs
@@ -5,19 +5,28 @@ use std::sync::Arc;
 use serde::{Deserialize, Serialize};
 
 use symbolicator_service::caching::CacheError;
-use symbolicator_service::types::{RawObjectInfo, Scope, ScrapingConfig};
+use symbolicator_service::types::{Platform, Scope, ScrapingConfig};
 use symbolicator_sources::{SentryFileId, SentrySourceConfig};
 
 use crate::lookup::CachedFileUri;
 
 #[derive(Debug, Clone)]
 pub struct SymbolicateJsStacktraces {
+    /// The event's platform.
+    ///
+    /// `Platform` is actually too lenient of a type here—a legitimate
+    /// JS event should only have
+    /// [`JavaScript`](symbolicator_service::types::JsPlatform::JavaScript)
+    /// or [`Node`](symbolicator_service::types::JsPlatform::Node). However,
+    /// we use the general `Platform` type for now to be resilient against
+    /// wrong values making it through.
+    pub platform: Option<Platform>,
     pub scope: Scope,
     pub source: Arc<SentrySourceConfig>,
     pub release: Option<String>,
     pub dist: Option<String>,
     pub stacktraces: Vec<JsStacktrace>,
-    pub modules: Vec<RawObjectInfo>,
+    pub modules: Vec<JsModule>,
     pub scraping: ScrapingConfig,
     /// Whether to apply source context for the stack frames.
     pub apply_source_context: bool,
@@ -75,6 +84,31 @@ pub struct JsModuleError {
     pub abs_path: String,
     #[serde(flatten)]
     pub kind: JsModuleErrorKind,
+}
+
+/// A JS module (representing a minified file and sourcemap).
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, PartialOrd, Ord)]
+pub struct JsModule {
+    /// The type of the module.
+    ///
+    /// This field carries no useful information because `JsModuleType`
+    /// only has one variant. It is used to distinguish `JsModule`s from
+    /// JVM and native modules during deserialization.
+    pub r#type: JsModuleType,
+    /// The module's minified code file.
+    pub code_file: String,
+    /// The module's debug ID.
+    pub debug_id: String,
+}
+
+/// The type of a sourcemap module.
+///
+/// As per <https://develop.sentry.dev/sdk/data-model/event-payloads/debugmeta/#source-map-images>,
+/// this is always `"sourcemap"`.
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq, PartialOrd, Ord)]
+#[serde(rename_all = "snake_case")]
+pub enum JsModuleType {
+    Sourcemap,
 }
 
 /// An attempt to scrape a JS source or sourcemap file from the web.
@@ -142,6 +176,7 @@ impl From<CacheError> for JsScrapingResult {
             ),
             CacheError::DownloadError(details) => (JsScrapingFailureReason::DownloadError, details),
             CacheError::Malformed(details) => (JsScrapingFailureReason::Other, details),
+            CacheError::Unsupported(details) => (JsScrapingFailureReason::Other, details),
             CacheError::InternalError => (JsScrapingFailureReason::Other, String::new()),
         };
 
@@ -174,6 +209,17 @@ pub enum JsScrapingFailureReason {
 
 #[derive(Debug, Default, Clone, Deserialize, Serialize, PartialEq, Eq)]
 pub struct JsFrame {
+    /// The frame's platform.
+    ///
+    /// `Platform` is actually too lenient of a type here—a legitimate
+    /// JS frame should only have
+    /// [`JavaScript`](symbolicator_service::types::JsPlatform::JavaScript)
+    /// or [`Node`](symbolicator_service::types::JsPlatform::Node). However,
+    /// we use the general `Platform` type for now to be resilient against
+    /// wrong values making it through.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub platform: Option<Platform>,
+
     #[serde(skip_serializing_if = "Option::is_none")]
     pub function: Option<String>,