+++ crates/symbolicator-proguard/src/symbolication.rs
@@ -4,6 +4,7 @@ use crate::interface::{
     CompletedJvmSymbolicationResponse, JvmException, JvmFrame, JvmModuleType, JvmStacktrace,
     ProguardError, ProguardErrorKind, SymbolicateJvmStacktraces,
 };
+use crate::metrics::{record_symbolication_metrics, SymbolicationStats};
 use crate::ProguardService;
 
 use futures::future;
@@ -28,6 +29,7 @@ impl ProguardService {
         request: SymbolicateJvmStacktraces,
     ) -> CompletedJvmSymbolicationResponse {
         let SymbolicateJvmStacktraces {
+            platform,
             scope,
             sources,
             exceptions,
@@ -37,6 +39,8 @@ impl ProguardService {
             apply_source_context,
             classes,
         } = request;
+
+        let mut stats = SymbolicationStats::default();
 
         let maybe_mappers = future::join_all(
             modules
@@ -112,11 +116,20 @@ impl ProguardService {
             })
             .collect();
 
-        let remapped_exceptions = exceptions
+        let remapped_exceptions: Vec<_> = exceptions
             .into_iter()
-            .map(|raw_exception| {
-                Self::map_exception(&mappers, &raw_exception).unwrap_or(raw_exception)
-            })
+            .map(
+                |raw_exception| match Self::map_exception(&mappers, &raw_exception) {
+                    Some(exc) => {
+                        stats.symbolicated_exceptions += 1;
+                        exc
+                    }
+                    None => {
+                        stats.unsymbolicated_exceptions += 1;
+                        raw_exception
+                    }
+                },
+            )
             .collect();
 
         let mut remapped_stacktraces: Vec<_> = stacktraces
@@ -126,7 +139,8 @@ impl ProguardService {
                     .frames
                     .iter()
                     .flat_map(|frame| {
-                        Self::map_frame(&mappers, frame, release_package.as_deref()).into_iter()
+                        Self::map_frame(&mappers, frame, release_package.as_deref(), &mut stats)
+                            .into_iter()
                     })
                     .collect();
                 JvmStacktrace {
@@ -147,12 +161,22 @@ impl ProguardService {
         let remapped_classes = classes
             .into_iter()
             .filter_map(|class| {
-                let remapped = mappers
-                    .iter()
-                    .find_map(|mapper| mapper.remap_class(&class))?;
-                Some((class, Arc::from(remapped)))
+                match mappers.iter().find_map(|mapper| mapper.remap_class(&class)) {
+                    Some(remapped) => {
+                        stats.symbolicated_classes += 1;
+                        Some((class, Arc::from(remapped)))
+                    }
+                    None => {
+                        stats.unsymbolicated_classes += 1;
+                        None
+                    }
+                }
             })
             .collect();
+
+        stats.num_stacktraces = remapped_stacktraces.len() as u64;
+
+        record_symbolication_metrics(platform, stats);
 
         CompletedJvmSymbolicationResponse {
             exceptions: remapped_exceptions,
@@ -207,6 +231,7 @@ impl ProguardService {
         mappers: &[&proguard::ProguardCache],
         frame: &JvmFrame,
         release_package: Option<&str>,
+        stats: &mut SymbolicationStats,
     ) -> Vec<JvmFrame> {
         let deobfuscated_signature = frame.signature.as_ref().and_then(|signature| {
             mappers
@@ -271,7 +296,7 @@ impl ProguardService {
 
         // Fix up the frames' in-app fields only if they were actually mapped
         if let Some(frames) = frames.as_mut() {
-            for frame in frames {
+            for frame in frames.iter_mut() {
                 // mark the frame as in_app after deobfuscation based on the release package name
                 // only if it's not present
                 if let Some(package) = release_package {
@@ -280,10 +305,24 @@ impl ProguardService {
                     }
                 }
             }
+
+            // Also count the frames as symbolicated at this point
+            for frame in frames {
+                *stats
+                    .symbolicated_frames
+                    .entry(frame.platform.clone())
+                    .or_default() += 1;
+            }
         }
 
         // If all else fails, just return the original frame.
-        let mut frames = frames.unwrap_or_else(|| vec![frame.clone()]);
+        let mut frames = frames.unwrap_or_else(|| {
+            *stats
+                .unsymbolicated_frames
+                .entry(frame.platform.clone())
+                .or_default() += 1;
+            vec![frame.clone()]
+        });
 
         for frame in &mut frames {
             // add the signature if we received one and we were
@@ -558,7 +597,10 @@ io.sentry.sample.MainActivity -> io.sentry.sample.MainActivity:
 
         let mapped_frames: Vec<_> = frames
             .iter()
-            .flat_map(|frame| ProguardService::map_frame(&[&cache], frame, None).into_iter())
+            .flat_map(|frame| {
+                ProguardService::map_frame(&[&cache], frame, None, &mut Default::default())
+                    .into_iter()
+            })
             .collect();
 
         assert_eq!(mapped_frames.len(), 7);
@@ -668,7 +710,13 @@ org.slf4j.helpers.Util$ClassContext -> org.a.b.g$b:
         let mapped_frames: Vec<_> = frames
             .iter()
             .flat_map(|frame| {
-                ProguardService::map_frame(&[&cache], frame, Some("org.slf4j")).into_iter()
+                ProguardService::map_frame(
+                    &[&cache],
+                    frame,
+                    Some("org.slf4j"),
+                    &mut Default::default(),
+                )
+                .into_iter()
             })
             .collect();
 
@@ -692,7 +740,8 @@ org.slf4j.helpers.Util$ClassContext -> org.a.b.g$b:
             ..Default::default()
         };
 
-        let remapped = ProguardService::map_frame(&[], &frame, Some("android"));
+        let remapped =
+            ProguardService::map_frame(&[], &frame, Some("android"), &mut Default::default());
 
         assert_eq!(remapped.len(), 1);
         // The frame didn't get mapped, so we shouldn't set `in_app` even though
@@ -739,7 +788,8 @@ org.slf4j.helpers.Util$ClassContext -> org.a.b.g$b:
             ..Default::default()
         };
 
-        let mapped_frames = ProguardService::map_frame(&[&cache], &frame, None);
+        let mapped_frames =
+            ProguardService::map_frame(&[&cache], &frame, None, &mut Default::default());
 
         assert_eq!(mapped_frames.len(), 2);
 
@@ -809,7 +859,8 @@ y.b -> y.b:
             ..Default::default()
         };
 
-        let mapped_frames = ProguardService::map_frame(&[&cache], &frame, None);
+        let mapped_frames =
+            ProguardService::map_frame(&[&cache], &frame, None, &mut Default::default());
 
         assert_eq!(mapped_frames.len(), 1);
 
@@ -948,7 +999,10 @@ io.wzieba.r8fullmoderenamessources.R -> a.d:
 
         let (remapped_filenames, remapped_abs_paths): (Vec<_>, Vec<_>) = frames
             .iter()
-            .flat_map(|frame| ProguardService::map_frame(&[&cache], frame, None).into_iter())
+            .flat_map(|frame| {
+                ProguardService::map_frame(&[&cache], frame, None, &mut Default::default())
+                    .into_iter()
+            })
             .map(|frame| (frame.filename.unwrap(), frame.abs_path.unwrap()))
             .unzip();